+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/simple/TrigonometryTests.java
@@ -1097,23 +1097,23 @@ public class TrigonometryTests extends SqlIoTest {
 
     @Test
     public void invalidInputs() {
-        this.shouldFail(
+        this.queryFailingInCompilation(
                 "SELECT sin(CAST('2023-12-14' AS DATE))",
                 "Error in SQL statement: Cannot apply 'SIN' to arguments of type 'SIN(<DATE>)'. " +
                         "Supported form(s): 'SIN(<NUMERIC>)'"
         );
 
-        this.shouldFail("SELECT sin(true)",
+        this.queryFailingInCompilation("SELECT sin(true)",
                 "Error in SQL statement: Cannot apply 'SIN' to arguments of type 'SIN(<BOOLEAN>)'. " +
                         "Supported form(s): 'SIN(<NUMERIC>)'"
         );
 
-        this.shouldFail("SELECT sin(CAST('101' AS BINARY(3)))",
+        this.queryFailingInCompilation("SELECT sin(CAST('101' AS BINARY(3)))",
                 "Error in SQL statement: Cannot apply 'SIN' to arguments of type 'SIN(<BINARY(3)>)'. " +
                         "Supported form(s): 'SIN(<NUMERIC>)'"
         );
 
-        this.shouldFail("SELECT sin(CAST('15:06:51.06731' AS TIME))",
+        this.queryFailingInCompilation("SELECT sin(CAST('15:06:51.06731' AS TIME))",
                 "Error in SQL statement: Cannot apply 'SIN' to arguments of type 'SIN(<TIME(0)>)'. " +
                         "Supported form(s): 'SIN(<NUMERIC>)'"
         );